requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
laspy[lazrs]>=0.7.0
rasterio>=1.2.0
pyproj>=3.2.0
pytest>=7.0.0
matplotlib>=3.5.0
pyyaml>=6.0
tqdm>=4.62.0
